refactor: move showFile to separate file
